// packages/appkit-react/src/features/balances/components/balance-badge/balance-badge.d.ts
/**
 * Copyright (c) TonTech.
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root directory of this source tree.
 *
 */
import type { FC } from 'react';
export declare const BalanceBadge: {
    Container: FC<import("react").DetailedHTMLProps<import("react").HTMLAttributes<HTMLDivElement>, HTMLDivElement>>;
    Icon: FC<import("../../../..").CircleIconProps>;
    BalanceBlock: FC<import("react").DetailedHTMLProps<import("react").HTMLAttributes<HTMLDivElement>, HTMLDivElement>>;
    Symbol: FC<import("react").ClassAttributes<HTMLSpanElement> & import("react").HTMLAttributes<HTMLSpanElement> & {
        symbol: string;
    }>;
    Balance: FC<import("react").ClassAttributes<HTMLSpanElement> & import("react").HTMLAttributes<HTMLSpanElement> & {
        balance: string;
        decimals: number;
    }>;
};
//# sourceMappingURL=balance-badge.d.ts.map